pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[project]
name = "scnn"
authors = [
    {name = "Song Kim", email = "songk@mit.edu"},
    {name = "Tuong Phung", email = "geiger.mario@gmail.com"},
]
description = "Comparison of convolution operations for image classification"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["cv"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy",
    "absl-py",
    "torch",
    "wandb"
]
dynamic = ["version"]

[tool.setuptools.packages]
find = {}

[tool.black]
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | .nox
)/
'''
